refactor: relocate shared UI state and BaseViewModel classes
