+++ ui/src/main/java/com/example/ui/navigation/NavGraph.kt
@@ -1,11 +1,8 @@
 package com.example.ui.navigation
 
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
-import androidx.navigation.compose.composable
 import com.example.ui.screens.categories.categoriesScreenRoute
 import com.example.ui.screens.home.homeScreenRoute
 import com.example.ui.screens.letsPlay.letsPlayScreenRoute
@@ -13,7 +10,7 @@ import com.example.ui.screens.lists.listsScreenRoute
 import com.example.ui.screens.profile.profileScreenRoute
 import com.example.ui.screens.search.searchScreenRoute
 import com.example.ui.screens.searchByActor.searchByActorScreenRoute
-import com.example.ui.screens.search.searchByCountry.searchByCountryScreenRoute
+import com.example.ui.screens.searchByCountry.searchByCountryScreenRoute
 
 @Composable
 fun NavGraph(
