+++ ui/src/main/java/com/example/ui/screens/searchByCountry/SearchByCountryScreenRoute.kt
@@ -1,4 +1,4 @@
-package com.example.ui.screens.search.searchByCountry
+package com.example.ui.screens.searchByCountry
 
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.compose.composable
