+++ viewModel/src/main/java/com/example/viewmodel/search/mapper/MovieItemUiStateMapper.kt
@@ -2,7 +2,7 @@ package com.example.viewmodel.search.mapper
 
 import android.icu.text.DecimalFormat
 import com.example.entity.Movie
-import com.example.viewmodel.shared.MovieItemUiState
+import com.example.viewmodel.shared.uiStates.MovieItemUiState
 
 private fun Movie.toMediaItemUiState(): MovieItemUiState =
     MovieItemUiState(
